Basic example showing all functions of the DS3132RTC
This sketch uses the Wire.h library for fast-mode 400 kHz I2C communications. The sketch runs on a 3.3V 8MHz Pro Mini ATMEGA 328P microprocessor, but should work on most other Arduino-compatible systems. The hardware interrupt 0 is pin #2 on the Pro Mini; it might have to changed if using a different ATMEGA chip.