Added EEPROM read/write functions
Added capability to read to and write from the 4k bytes of EEPROM so can record alarm sounds and gongs for the hour and alarms, etc. Demonstrate alarm sounds every 30 seconds as a test of the capability. Next stop is to design a clock with proper hourly and alarm sound responses.